@@ -1,18 +1,24 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+    <WarningsAsErrors />
+    <LangVersion>latest</LangVersion>
+    <!--<AnalysisMode>AllEnabledByDefault</AnalysisMode>-->
   </PropertyGroup>
+
   <ItemGroup>
-    <PackageReference Include="AWSSDK.S3" Version="3.3.102.12"/>
-    <PackageReference Include="AWSSDK.Extensions.NETCore.Setup" Version="3.3.100.1"/>
-    <PackageReference Include="Microsoft.Extensions.Configuration" Version="2.2.0"/>
-    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="2.2.0"/>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="2.2.0"/>
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="2.2.0"/>
-    <PackageReference Include="Microsoft.Extensions.Logging.Configuration" Version="2.2.0"/>
-    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="2.2.0"/>
-    <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="2.2.0"/>
-    <PackageReference Include="Json.Net" Version="1.0.16"/>
+    <PackageReference Include="AWSSDK.S3" Version="3.7.0.8" />
+    <PackageReference Include="AWSSDK.Extensions.NETCore.Setup" Version="3.7.0.1" />
+    <PackageReference Include="Microsoft.Extensions.Configuration" Version="5.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="5.0.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="5.0.1" />
+    <PackageReference Include="Microsoft.Extensions.Logging" Version="5.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Configuration" Version="5.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="5.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="5.0.0" />
+    <PackageReference Include="Json.Net" Version="1.0.25" />
   </ItemGroup>
 </Project>
