@@ -1,27 +1,19 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.HttpsPolicy;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using System;
 
 namespace Cognito.MvcApp
 {
     public class Startup
     {
-        public Startup(IConfiguration configuration, IHostingEnvironment env)
-        {
-            Configuration = configuration;
-            Env = env;
-        }
+        private readonly IWebHostEnvironment _environment;
 
-        private IConfiguration Configuration { get; }
-        private IHostingEnvironment Env { get; }
+        public Startup(IWebHostEnvironment env)
+        {
+            _environment = env ?? throw new ArgumentNullException(nameof(env));
+        }
 
         public void ConfigureServices(IServiceCollection services)
         {
@@ -30,13 +22,13 @@ namespace Cognito.MvcApp
             {
                 options.LoginPath = "/Accounts/SignIn";
             });
-            
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+
+            services.AddControllersWithViews();
         }
 
         public void Configure(IApplicationBuilder app)
         {
-            if (Env.IsDevelopment())
+            if (_environment.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
@@ -47,17 +39,18 @@ namespace Cognito.MvcApp
             }
 
             app.UseStaticFiles();
-            app.UseNodeModules(Env);
-            
+            app.UseNodeModules();
+
             const string PathFormat = "/errors/{0}";
             app.UseStatusCodePagesWithReExecute(PathFormat);
-
             app.UseAuthentication();
-            app.UseMvc(routes =>
+            app.UseRouting();
+            app.UseAuthorization();
+            app.UseEndpoints(endpoints =>
             {
-                routes.MapRoute(
+                endpoints.MapControllerRoute(
                     name: "default",
-                    template: "{controller=Home}/{action=Index}/{id?}");
+                    pattern: "{controller=Home}/{action=Index}/{id?}");
             });
         }
     }
