@@ -8,15 +8,15 @@ namespace Cognito.MvcApi.Models.Signups
     {
         [Required(ErrorMessage = "An email address is required")]
         [DataType(DataType.EmailAddress)]
-        public string Email { get; set; }
+        public string Email { get; set; } = string.Empty;
 
         [Required(ErrorMessage = "A password is required")]
         [DataType(DataType.Password)]
-        public string Password { get; set; }
+        public string Password { get; set; } = string.Empty;
 
         [Required(ErrorMessage = "A confirmed password is required")]
         [Compare(nameof(Password), ErrorMessage = "Passwords must match")]
         [DataType(DataType.Password)]
-        public string ConfirmedPassword { get; set; }        
+        public string ConfirmedPassword { get; set; } = string.Empty;
     }
 }
