@@ -8,10 +8,10 @@ namespace Cognito.MvcApi.Models.Tokens
     {
         [Required(ErrorMessage = "An email address is required")]
         [DataType(DataType.EmailAddress)]
-        public string Email { get; set; }
+        public string Email { get; set; } = string.Empty;
 
         [Required(ErrorMessage = "A password is required")]
         [DataType(DataType.Password)]
-        public string Password { get; set; }       
+        public string Password { get; set; } = string.Empty;
     }
 }
