@@ -7,11 +7,11 @@ namespace Cognito.MvcApp.ViewModels.Accounts
         [Required(ErrorMessage = "An email is required")]
         [DataType(DataType.EmailAddress)]
         [Display(Name = "Email")]
-        public string Email { get; set; }
+        public string Email { get; set; } = string.Empty;
 
         [Required(ErrorMessage = "A password is required")]
         [DataType(DataType.Password)]
         [Display(Name = "Password")]
-        public string Password { get; set; }
+        public string Password { get; set; } = string.Empty;
     }
 }
