@@ -1,5 +1,4 @@
 @using Microsoft.AspNetCore.Identity
-@using Microsoft.AspNetCore.Identity.EntityFrameworkCore
 @using Amazon.Extensions.CognitoAuthentication
 
 @inject UserManager<CognitoUser> UserManager
@@ -14,42 +13,42 @@
 </head>
 <body>
     <header class="container">
-        <nav class="navbar navbar-expand-lg navbar-light bg-light">         
+        <nav class="navbar navbar-expand-lg navbar-light bg-light">
             <a asp-controller="Notes" asp-action="Index" class="navbar-brand text-primary">
-            <em class="fa fa-book-open fa-2x mr-2 text-primary"></em>
-            COGNITO
+                <em class="fa fa-book-open fa-2x mr-2 text-primary"></em>
+                COGNITO
             </a>
             <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#menu">
-            <span class="navbar-toggler-icon"></span>
+                <span class="navbar-toggler-icon"></span>
             </button>
 
             <div class="collapse navbar-collapse" id="menu">
-            <form asp-controller="Accounts" asp-action="SignOut" class="my-2 ml-auto my-lg-0">                
+                <form asp-controller="Accounts" asp-action="SignOut" class="my-2 ml-auto my-lg-0">
                     <ul class="navbar-nav">
                         <li class="nav-item dropdown">
-                        <a class="nav-link dropdown-toggle" href="#" id="accountDropdown" role="button" data-toggle="dropdown">
-                            <div class="d-inline d-lg-none">ACCOUNT</div>
-                            <div class="d-none d-lg-inline">
-                                <span class="fa-stack">
-                                <em class="fas fa-circle fa-stack-2x text-primary"></em>
-                                <em class="fas fa-user fa-stack-1x fa-inverse"></em>
-                                </span>
-                                @UserManager.GetUserName(User)
-                            </div>                                
-                        </a>
-                        <div class="dropdown-menu">
-                            <button type="submit" class="dropdown-item btn btn-link">Logout</button>
-                        </div>                        
+                            <a class="nav-link dropdown-toggle" href="#" id="accountDropdown" role="button" data-toggle="dropdown">
+                                <div class="d-inline d-lg-none">ACCOUNT</div>
+                                <div class="d-none d-lg-inline">
+                                    <span class="fa-stack">
+                                        <em class="fas fa-circle fa-stack-2x text-primary"></em>
+                                        <em class="fas fa-user fa-stack-1x fa-inverse"></em>
+                                    </span>
+                                    @UserManager.GetUserName(User)
+                                </div>
+                            </a>
+                            <div class="dropdown-menu">
+                                <button type="submit" class="dropdown-item btn btn-link">Logout</button>
+                            </div>
                         </li>
                     </ul>
-            </form>
+                </form>
             </div>
         </nav>
     </header>
     <main class="container">
         @RenderBody()
     </main>
-    <footer></footer>    
+    <footer></footer>
     @await Html.PartialAsync("_Scripts")
     @RenderSection("Scripts", required: false)
 </body>
