@@ -6,12 +6,12 @@ namespace DynamoDb.ConsoleApp.Repositories.Models
     public sealed class BookEntity
     {
         [DynamoDBHashKey]
-        public string Id { get; set; }
+        public string Id { get; set; } = string.Empty;
 
         [DynamoDBProperty]
-        public string Title { get; set; }
+        public string Title { get; set; } = string.Empty;
 
         [DynamoDBProperty]
-        public string Description { get; set; }
+        public string Description { get; set; } = string.Empty;
     }
 }
