@@ -5,7 +5,7 @@ namespace DynamoDb.ConsoleApp.Services.Books
     [Serializable]
     public sealed class BookForCreate
     {
-        public string Title { get; set; }
-        public string Description { get; set; }
+        public string Title { get; set; } = string.Empty;
+        public string Description { get; set; } = string.Empty;
     }
 }
