@@ -2,6 +2,6 @@ namespace S3.ConsoleApp.Services
 {
     public sealed class DeletedBucket
     {
-        public string Name { get; set; }
+        public string Name { get; set; } = string.Empty;
     }
 }
