@@ -2,6 +2,6 @@ namespace Sns.ConsoleApp.Services.Publications
 {
     public sealed class PublishConfirmation
     {
-        public string MessageId { get; set; }        
+        public string MessageId { get; set; } = string.Empty;
     }
 }
