@@ -2,6 +2,6 @@ namespace Sns.ConsoleApp.Services.Subscriptions
 {
     public sealed class SubscriptionCancellation
     {
-        public string SubscriptionArn { get; set; }
+        public string SubscriptionArn { get; set; } = string.Empty;
     }
 }
