@@ -5,9 +5,9 @@ namespace Sns.ConsoleApp.Services.Topics
 {
     public interface ITopicService
     {
-        Task<string> CreateTopicAsync(string topicName, IDictionary<string, string> attributes = null);
+        Task<string> CreateTopicAsync(string topicName, IDictionary<string, string>? attributes = null);
         Task DeleteTopicAsync(string topicName);
         Task<IReadOnlyList<TopicDetail>> GetAllTopicsAsync();
-        Task<TopicDetail> GetTopicAsync(string topicName);
+        Task<TopicDetail?> GetTopicAsync(string topicName);
     }
 }
