@@ -4,8 +4,8 @@ namespace Sns.ConsoleApp.Services.Topics
 {
     public sealed class TopicDetail
     {
-        public string Name { get; set; }
-        public string Region { get; set; }
-        public IReadOnlyDictionary<string, string> Attributes { get; set; }
+        public string Name { get; set; } = string.Empty;
+        public string Region { get; set; } = string.Empty;
+        public IReadOnlyDictionary<string, string> Attributes { get; set; } = new Dictionary<string, string>();
     }
 }
